Extract high-value style boundary policy
